requirements.txt
gradio>=6.2.0
mcp>=1.25.0
mistralai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
